refactor: move general function outside of the result store
